refactor(send): PR 2 phase 2c+2d+2e — full SendDetailsScreen migration + delete legacy
Completes Phase 2 of the form-VM rewrite. Atomic commit that:

1. Renames `SendDetailsFormViewModel` → `SendDetailsViewModel` (file + class).
   Tests + fixture refs updated.
2. Deletes the old UI-state-only `SendDetailsViewModel` (its content was
   already merged into the form VM in phase 2a).
3. Rewrites `SendDetailsScreen` to use a single `@State viewModel:
   SendDetailsViewModel` replacing the three legacy state holders
   (`@ObservedObject tx: LegacySendTransaction`, `@StateObject sendCryptoViewModel`,
   `@StateObject sendDetailsViewModel`). All ~40 `tx.X` reads, ~30
   `sendCryptoViewModel.X` calls, and ~20 `sendDetailsViewModel.X` calls
   collapse to `viewModel.X`.
4. Rewrites the 7 `SendDetails*` components (`AssetTab`, `AddressTab`,
   `AddressFields`, `AmountTab`, `AmountTextField`, `AdditionalSection`,
   `TabEditTools`) to take `@Bindable var viewModel: SendDetailsViewModel`
   instead of the legacy trio. SwiftUI binding pattern switches from
   `$tx.X` → `$viewModel.X`.
5. Deletes `SendCryptoViewModel.swift` (518 LOC) + `LegacySendCryptoInteractor`.
   3 production consumers outside Send Details (`TronFreezeView`,
   `TronUnfreezeView`, `CircleDepositViewModel`) replaced their
   `SendCryptoViewModel()` with `DefaultSendInteractor.live` for the only
   call site they used (`loadFastVault`).
6. Adds `@MainActor` to 3 view builders (`SendRouteBuilder.buildDetailsScreen`,
   `SendRouter.build`, `HomeRouter.build`, `HomeRouteBuilder.buildActionRoute`)
   since the new VM's init is MainActor-isolated.

Behaviour changes that should be smoke-tested:
- Sheet presentation paths (`SendGasSettingsView`, chain picker, coin picker)
  now bind through `$viewModel.X` instead of `$tx.X` + `$sendDetailsViewModel.X`.
- Address resolution: `viewModel.debouncedResolveAddress()` replaces the
  old `sendCryptoViewModel.debouncedResolveAddress(tx:)` — same shape,
  just the new VM owns the task.
- Percentage buttons now call `Task { await viewModel.setMaxAmount(percentage:) }`
  instead of `sendCryptoViewModel.setMaxValues(tx:percentage:)`. Per-chain
  logic delegated to `SendInteractor`.
- `tx.reset(coin:)` in `setMainData` becomes `viewModel.reset(to:)`.
- The fast-vault check on TRON/Circle paths goes through the new
  `DefaultSendInteractor.loadFastVault` which includes the explicit
  `hasPrefix("server-")` guard (Phase D lesson).

108 existing tests still pass. Build green. Component-level SwiftUI
binding regressions require manual simulator smoke (Details → Verify
happy paths on UTXO/EVM/Cosmos, sheet presentations, address resolver,
custom-gas pinning).
